pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ffmpeg4discord"
version = "0.1.2"
description = "A tool to help convert a video file to a target file size."
readme = {file = "README.md", content-type = "text/markdown"}
authors = [{ name = "Zach Fleeman", email = "zfleeman@gmail.com" }]
requires-python = ">= 3.8"
dependencies = [
    "Flask==3.0.1",
    "ffmpeg-python==0.2.0"
]
classifiers = [
    "License :: OSI Approved :: MIT License",
]

[project.urls]
Homepage = "https://github.com/zfleeman/ffmpeg4discord"
Issues = "https://github.com/zfleeman/ffmpeg4discord/issues"

[project.scripts]
ff4d = "ffmpeg4discord.__main__:main"
install-ffmpeg-windows = "ffmpeg4discord.windows:install"

[tool.setuptools.package-data]
ffmpeg4discord = ["templates/*.html"]
